Remove unused chunks and save shadow coord generation code for clustered lights.
