+++ src/scene/shader-lib/chunks/chunks.js
@@ -150,9 +150,6 @@ import refractionDynamicPS from './lit/frag/refractionDynamic.js';
 import reprojectPS from './common/frag/reproject.js';
 import screenDepthPS from './common/frag/screenDepth.js';
 import shadowCascadesPS from './lit/frag/shadowCascades.js';
-import shadowCommonPS from './lit/frag/shadowCommon.js';
-import shadowCoordPS from './lit/frag/shadowCoord.js';
-import shadowCoordPerspZbufferPS from './lit/frag/shadowCoordPerspZbuffer.js';
 import shadowEVSMPS from './lit/frag/shadowEVSM.js';
 import shadowEVSMnPS from './lit/frag/shadowEVSMn.js';
 import shadowSampleCoordPS from './lit/frag/shadowSampleCoord.js';
@@ -359,9 +356,6 @@ const shaderChunks = {
     reprojectPS,
     screenDepthPS,
     shadowCascadesPS,
-    shadowCommonPS,
-    shadowCoordPS,
-    shadowCoordPerspZbufferPS,
     shadowEVSMPS,
     shadowEVSMnPS,
     shadowSampleCoordPS,
