--- src/scene/shader-lib/chunks/lit/frag/shadowCoordPerspZbuffer.js
@@ -1,17 +1,0 @@
-export default /* glsl */`
-void _getShadowCoordPerspZbuffer(mat4 shadowMatrix, vec4 shadowParams, vec3 wPos) {
-    vec4 projPos = shadowMatrix * vec4(wPos, 1.0);
-    projPos.xyz /= projPos.w;
-    dShadowCoord = projPos.xyz;
-    // depth bias is already applied on render
-}
-
-void getShadowCoordPerspZbufferNormalOffset(mat4 shadowMatrix, vec4 shadowParams, vec3 normal) {
-    vec3 wPos = vPositionW + normal * shadowParams.y;
-    _getShadowCoordPerspZbuffer(shadowMatrix, shadowParams, wPos);
-}
-
-void getShadowCoordPerspZbuffer(mat4 shadowMatrix, vec4 shadowParams) {
-    _getShadowCoordPerspZbuffer(shadowMatrix, shadowParams, vPositionW);
-}
-`;
